fixed the review section
